add editPoolConfig; allow editing maker broker (#89)
* add editPoolConfig; allow editing maker broker

* fix rust client and tests

* regen idl

* enhance rust client

* update to use validate_core_asset from toolbox

+++ clients/rust/src/hooked/mod.rs
@@ -7,7 +7,7 @@ pub use currency::*;
 pub use fees::*;
 pub use nullable::*;
 
-pub const HUNDRED_PCT_BPS: u16 = 10000;
+pub const HUNDRED_PCT_BPS: u64 = 10000;
 pub const LAMPORTS_PER_SOL: u64 = 1_000_000_000;
 pub const BROKER_FEE_PCT: u64 = 50;
 pub const MAKER_BROKER_PCT: u8 = 80;
